requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
clikit==0.6.2
cmdstanpy==1.0.4
commonmark==0.9.1
convertdate==2.4.0
crashtest==0.3.1
cycler==0.11.0
Cython==0.29.30
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dukpy==0.2.3
entrypoints==0.4
ephem==4.1.3
executing==0.9.1
fastjsonschema==2.16.1
filelock==3.7.1
fonttools==4.34.4
frozenlist==1.3.0
gitdb==4.0.9
GitPython==3.1.27
hijri-converter==2.2.4
holidays==0.14.2
httpstan==4.8.0
idna==3.3
importlib-metadata==4.12.0
install==1.3.5
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
korean-lunar-calendar==0.2.1
localpip==1.0.7
LunarCalendar==0.0.9
lxml==4.9.1
MarkupSafe==2.1.1
marshmallow==3.17.0
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
multitasking==0.0.11
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.1
packaging==21.3
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.3
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.9.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
prophet==1.1
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.12.0
pylev==1.4.0
PyMeeus==0.5.11
Pympler==1.0.1
pyobjc-core==8.5
pyobjc-framework-Cocoa==8.5
pyobjc-framework-SystemConfiguration==8.5
PyPAC==0.16.0
pyparsing==3.0.9
pyrsistent==0.18.1
pysimdjson==3.2.0
pystan==2.19.1.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.0
requests==2.26.0
requests-file==1.5.1
rich==12.5.1
scikit-learn==1.1.1
scipy==1.9.0
semver==2.13.0
Send2Trash==1.8.0
setuptools-git==1.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.3.0
streamlit==1.11.0
tenacity==8.0.1
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tldextract==3.3.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
ujson==5.4.0
urllib3==1.26.11
validators==0.20.0
wcwidth==0.2.5
webargs==8.2.0
webencodings==0.5.1
widgetsnbextension==3.6.1
yarl==1.7.2
yfinance==0.1.74
zipp==3.8.1
